Moved Integration Test annotation
Also silenced more Validator warnings (which aren't actually warnings).
