+++ dpc-aggregation/src/test/java/gov/cms/dpc/aggregation/AggregationServiceTest.java
@@ -2,11 +2,11 @@ package gov.cms.dpc.aggregation;
 
 import com.codahale.metrics.health.HealthCheckRegistry;
 import gov.cms.dpc.testing.BufferedLoggerHandler;
+import gov.cms.dpc.testing.IntegrationTest;
 import io.dropwizard.testing.ConfigOverride;
 import io.dropwizard.testing.DropwizardTestSupport;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -15,8 +15,7 @@ import java.util.SortedSet;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-// Duplicating this tag so that we don't have to pull in more test resources or refactor into a shared package.
-@Tag("Integration")
+@IntegrationTest
 @ExtendWith(BufferedLoggerHandler.class)
 public class AggregationServiceTest {
