+++ dpc-testing/src/main/java/gov/cms/dpc/testing/IntegrationTest.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.api.annotations;
+package gov.cms.dpc.testing;
 
 import org.junit.jupiter.api.Tag;
